Update and rename General Breaks.ts to Core Breaks.ts
